+++ EssentialApp/EssentialApp/FeedImageDataLoaderWithFallbackComposite.swift
@@ -1,0 +1,35 @@
+import EssentialDeveloper
+
+public final class FeedImageDataLoaderWithFallbackComposite: FeedImageDataLoader {
+    let primary: FeedImageDataLoader
+    let fallback: FeedImageDataLoader
+    
+    private final class Task: FeedImageDataLoaderTask {
+        var wrapped: FeedImageDataLoaderTask?
+        
+        func cancel() {
+            wrapped?.cancel()
+        }
+    }
+    
+    public init(primary: FeedImageDataLoader, fallback: FeedImageDataLoader) {
+        self.primary = primary
+        self.fallback = fallback
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = Task()
+        
+        task.wrapped = primary.loadImageData(from: url) { [weak self] result in
+            switch result {
+            case .success:
+                completion(result)
+            case .failure:
+                task.wrapped = self?.fallback.loadImageData(from: url, completion: completion)
+            }
+        }
+        
+        
+        return task
+    }
+}
